feat: add functionality to upload and manage activity photos in reports

+++ src/main/java/com/branches/material/dto/request/CreateMaterialDeRelatorioRequest.java
@@ -3,11 +3,13 @@ package com.branches.material.dto.request;
 import com.branches.material.domain.enums.TipoMaterial;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record CreateMaterialDeRelatorioRequest(
         @NotNull(message = "O campo 'materialId' é obrigatório")
         Long materialId,
         String unidadeMedida,
-        Integer quantidade,
+        BigDecimal quantidade,
         @NotNull(message = "O campo 'tipoMaterial' é obrigatório")
         TipoMaterial tipoMaterial
 ) {
